+++ package.xml
@@ -10,6 +10,7 @@
   <license>Apache 2.0</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
+  <buildtool_depend>ament_cmake_python</buildtool_depend>
 
   <depend>tf2_eigen</depend>
   <depend>interactive_markers</depend>
